jigsaw to app conversion and remove unnecessary tests
