+++ isis/src/control/apps/jigsaw/jigsaw.h
@@ -1,0 +1,15 @@
+#ifndef jigsaw_h
+#define jigsaw_h
+
+#include <vector>
+
+#include <QString>
+
+#include "Pvl.h"
+#include "UserInterface.h"
+
+namespace Isis {
+  extern void jigsaw(UserInterface &ui, Pvl *log = nullptr);
+}
+
+#endif
